[#26] refactor : carousel 컴포넌트 생성
- 라이브러리화를 위해 carousel 컴포넌트 생성
