+++ frontend/src/components/organisms/Carousel/index.style.jsx
